+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `estimate_cost()` method for `Job` and `Batch`.
 - `s3utils.get_s3_sts_token` accepts extra query parameters.
 - `plugins.mode.web` to control the server-side mode solver.
+- Support for `JaxDataArray.interp`, allowing differentiable linear interpolation.
 
 ### Changed
 - Add `Medium2D` to full simulation in tests.